+++ src/applications/zxspectrum_48k/states/running_state.h
@@ -48,10 +48,10 @@ private:
     static constexpr unsigned int s_rst_7_z80 = 0xff;
 
     // Game loop - begin
-    static constexpr long double s_fps = 50.0L;
-    static constexpr long double s_tick_limit = 1000.0L / s_fps;
+    static constexpr double s_fps = 50.0;
+    static constexpr double s_tick_limit = 1000.0 / s_fps;
     static constexpr int s_cycles_per_ms = 3500;
-    static constexpr long double s_cycles_per_tick = s_cycles_per_ms * s_tick_limit;
+    static constexpr double s_cycles_per_tick = s_cycles_per_ms * s_tick_limit;
     // Game loop - end
 
     std::shared_ptr<StateContext> m_ctx;
